Реализовано: Авторизация, валидация запросов, обработчик exception-ов

+++ src/main/java/com/client/repository/ExtendedProfileRepository.java
@@ -3,5 +3,5 @@ package com.client.repository;
 import com.client.domain.ExtendedProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExtendedProfileRepository extends JpaRepository<ExtendedProfile, Integer> {
+public interface ExtendedProfileRepository extends JpaRepository<ExtendedProfile, Long> {
 }
